+++ wardenjs/src/codegen/cosmos/gov/module/v1/module.ts
@@ -102,7 +102,7 @@ export const Module = {
   },
   toAmino(message: Module): ModuleAmino {
     const obj: any = {};
-    obj.max_metadata_len = message.maxMetadataLen !== BigInt(0) ? message.maxMetadataLen.toString() : undefined;
+    obj.max_metadata_len = message.maxMetadataLen !== BigInt(0) ? (message.maxMetadataLen?.toString)() : undefined;
     obj.authority = message.authority === "" ? undefined : message.authority;
     return obj;
   },